Add Regal for linting Rego
Hi there, GKE friends 👋😃

This PR introduces [Regal](https://github.com/StyraInc/regal) for linting Rego policy
for the project. I've limited linting (and fixes) to the `gke-policies-v2` directory
here, but I'd be happy to expand this to the v1 directory as well if you think that'd
be valuable.

Along with a configuration file to disable some of the linters that'd require a more
concentrated effort to adhere to, I've fixed the following reported violations, which
while mostly style-related, seems like they'd be good to enforce for modern Rego:

* [use-some-for-output-vars](https://docs.styra.com/regal/rules/use-some-for-output-vars)
* [non-raw-regex-pattern](https://docs.styra.com/regal/rules/idiomatic/non-raw-regex-pattern)
* [prefer-package-imports](https://docs.styra.com/regal/rules/imports/prefer-package-imports)
* [chained-rule-body](https://docs.styra.com/regal/rules/style/chained-rule-body)
* [identically-named-tests](https://docs.styra.com/regal/rules/testing/identically-named-tests)
* [use-assignment-operator](https://docs.styra.com/regal/rules/style/use-assignment-operator)

I've also included linting as a step for the GitHub Action that runs tests of all the
policies.

Let me know what you think!

Signed-off-by: Anders Eknert <anders@styra.com>